// BankSystemUI/src/app/modules/clientes/interfaces/Clientes.model.ts
export interface CrearCliente{
    nombre: string;
    genero: string;
    edad: number;
    identificacion: number;
    direccion: string;
    telefono: number;
    contrasena: string;
    estado: boolean;
}

export interface Cliente{
    personaId: number;
    nombre: string;
    genero: string;
    edad: number;
    identificacion: number;
    direccion: string;
    telefono: number;
    contrasena: string;
    estado: boolean;
}